Move test helper file out of test_helpers
While you may think test_helpers is for helper files for test it is
actually used to unit test code under the GdsApi::TestHelpers namespace.

This moves the test helper code to be in a support directory to reflect
it is helper files and not unit tests.
